IMPALA-4957: Don't crash Impalad on LLVM fatal error
Whenever LLVM hits a fatal error condition, it would
call the registered callback LlvmCodegenHandleError()
which used to log the reason leading to the fatal
error and then trigger a core dump. The old behavior
is not desirable as codegen failure shouldn't lead
to Impalad crashing.

With this change, LlvmCodegenHandleError() will throw
an exception instead. a new try-catch block is also
added in LlvmCodeGen::FinalizeModule() to catch any
exceptions thrown and wind back the stack. This allows
fatal codegen error (e.g. unresolved symbols) to not
crash Impala. Instead, it just fails the query.

Change-Id: I54706e261ed223eadde347b1184fb0102e03a3d6