IMPALA-5558: Re-establish connection between ReportExecStatus RPC
QueryState::ReportExecStatusAux() aims to retry the RPC for up to
3 times when reporting exec status of a fragment instance to the
coordinator. However, it's not very effective because a new client
connection isn't established between retry. In addition, since
we are reporting the query profile, it should be fine to retry all
the time even if the payload was partially sent to the remote client.

This change fixes the problem by re-establishing the connection
before each RPC call. In addition, the code will unconditionally
retry the RPC up to 3 times.

Change-Id: I4d722c8ad3bf0e78e89887b6cb286c69ca61b8f5