IMPALA-6184: Handle ScalarExprEvaluator::Clone() failure properly
When ScalarExprEvaluator::Clone() fails, the newly created evaluator is not added
to the output vector. This makes it impossible for callers to close and clean up
the evaluators afterwards. This change fixes this by always adding the newly
created evaluator to the output vector before checking for the error status.

This path is only exercised in the scanner code. Two new tests are added to
exercise the failure paths.

Testing done: newly added tests in udf-errors.test

Change-Id: I45ffd722d0a69ad05ae3c748cf504c7f1a959a1d